backend/requirements.txt
langchain
langchain-community
langchain-google-genai
google-generativeai
fastapi
uvicorn
duckduckgo-search
python-dotenv
pydantic
